requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
plotly==5.24.1
streamlit==1.40.2
streamlit-autorefresh==1.0.1
python-telegram-bot==21.6